refactor(app): move the daemons screen state and actions into ainb-app
DaemonsState, its background collector and the action runner move to
ainb_app::components::daemons, along with what they call: `ainb daemon`
(cli::daemon) and the `fleet daemons` formatters (cli::fleet::daemons).
fleet::daemon_cta, which drives the same action runner, leaves the core
fleet shim.

The daemon verbs' argv test stays in ainb-core because it parses through
the full clap tree, which only that crate builds.
