+++ ainb-tui/crates/ainb-app/src/cli/fleet/mod.rs
@@ -1,0 +1,3 @@
+// ABOUTME: Renderer-agnostic `ainb fleet` pieces the daemons screen shares.
+
+pub mod daemons;
